move exceptions to common
